// apps/superdao-backend/src/typechain/DAOConstructor.ts
export type AdminSettingsStruct = {
  admins: string[];
  releaseManager: string;
  creator: string;
};

export type AdminSettingsStructOutput = [string[], string, string] & {
  admins: string[];
  releaseManager: string;
  creator: string;
};

export type AttributeStruct = {
  tierId: BytesLike;
  attrName: string;
  value: BytesLike;
};

export type AttributeStructOutput = [string, string, string] & {
  tierId: string;
  attrName: string;
  value: string;
};

export type NFTSettingsStruct = {
  openseaOwner: string;
  url: string;
  name: string;
  symbol: string;
  attributes: AttributeStruct[];
};

export type NFTSettingsStructOutput = [
  string,
  string,
  string,
  string,
  AttributeStructOutput[]
] & {
  openseaOwner: string;
  url: string;
  name: string;
  symbol: string;
  attributes: AttributeStructOutput[];
};

export type SaleSettingsStruct = {
  tiersValues: BytesLike[];
  tiersPrices: BigNumberish[];
  claimLimit: BigNumberish;
  tokenSaleAddress: string;
};

export type SaleSettingsStructOutput = [
  string[],
  BigNumber[],
  BigNumber,
  string
] & {
  tiersValues: string[];
  tiersPrices: BigNumber[];
  claimLimit: BigNumber;
  tokenSaleAddress: string;
};

export type DeploymentSettingsStruct = {
  adminSettings: AdminSettingsStruct;
  nftSettings: NFTSettingsStruct;
  openSaleSettings: SaleSettingsStruct;
  whiteListSaleSettings: SaleSettingsStruct;
};